# TitanFrontEnds/Titan.Client/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Titan.API",
    "description": "An ASP.NET Core Web API for Titan functions",
    "version": "v1"
  },
  "paths": {
...
    "/api/Orders/{OrderID}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "operationId": "GetOrderByID",
        "parameters": [
          {
            "name": "OrderID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the Order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDTO"
                }
              }
            }
          },
          "404": {
            "description": "Could not find the Order"
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "operationId": "UpdateOrder",
        "parameters": [
          {
            "name": "OrderID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succesfully order"
          },
          "400": {
            "description": "Cannot update order that is authorised"
          },
          "404": {
            "description": "Could not find the Order"
          }
        }
      }
    },
    "/api/Orders/{OrderID}/Details/{OrderDetailID}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "operationId": "GetOrderDetail",
        "parameters": [
          {
            "name": "OrderID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderDetailID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the Order detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Could not find Order or detail"
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "operationId": "UpdateOrderDetail",
        "parameters": [
          {
            "name": "OrderID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderDetailID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDetailDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDetailDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDetailDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns succesfully updated Order detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Could not find stock code"
          },
          "404": {
            "description": "Could not find Order or detail to update"
          }
        }
      }
    },
    "/api/Orders/{OrderID}/Details/Add": {
      "post": {
        "tags": [
          "Orders"
        ],
        "operationId": "AddOrderDetail",
        "parameters": [
          {
            "name": "OrderID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDetailDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDetailDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDetailDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns succesfully added Order detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Could not find Order"
          },
          "400": {
            "description": "Could not find stock code"
          }
        }
      }
    },
    "/api/Orders/{ConID}/PartsToBookIn": {
      "get": {
        "tags": [
          "Orders"
        ],
        "operationId": "GetOrderPartsSuggestions",
        "parameters": [
          {
            "name": "ConID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the Order detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReceivedPart"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Could not find contract"
          }
        }
      }
    },
    "/api/Orders/BookInParts": {
      "post": {
        "tags": [
          "Orders"
        ],
        "operationId": "BookInParts",
        "parameters": [
          {
            "name": "ConID",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "DeliveryRef",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "GRN",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ReceivedPart"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ReceivedPart"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ReceivedPart"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns succesfully added Order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDTO"
                }
              }
            }
          },
          "400": {
            "description": "Could not find contract to book in against"
          },
          "409": {
            "description": "Cannot book in a blank GRN"
          },
          "422": {
            "description": "Could not process a recieved part"
          }
        }
      }
    },
    "/api/Orders/{OrderID}/GenerateSOR": {
      "post": {
        "tags": [
          "Orders"
        ],
        "operationId": "ApproveAndGenerateSORForOrder",
        "parameters": [
          {
            "name": "OrderID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns succesfully created SOR number"
          },
          "400": {
            "description": "Cannot generate SOR for Order without customer"
          },
          "404": {
            "description": "Could not find Order to create SOR from"
          },
          "422": {
            "description": "Could not process order line"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrderDTO": {
        "type": "object",
        "properties": {
          "orderID": {
            "type": "integer",
            "format": "int32"
          },
          "conID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "dateReceived": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cusOrderNumber": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "cusID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ordCustomerRef": {
            "type": "string",
            "nullable": true
          },
          "ordSpecialInstruction": {
            "type": "string",
            "nullable": true
          },
          "postalName": {
            "type": "string",
            "nullable": true
          },
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "addressLine3": {
            "type": "string",
            "nullable": true
          },
          "addressLine4": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "county": {
            "type": "string",
            "nullable": true
          },
          "postCode": {
            "type": "string",
            "nullable": true
          },
          "deliveryTerms": {
            "type": "string",
            "nullable": true
          },
          "despatchMethod": {
            "type": "string",
            "nullable": true
          },
          "approved": {
            "type": "boolean"
          },
          "approvedBy": {
            "type": "string",
            "nullable": true
          },
          "sor": {
            "type": "string",
            "nullable": true
          },
          "grn": {
            "type": "string",
            "nullable": true
          },
          "orderDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDetailDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderDetailDTO": {
        "type": "object",
        "properties": {
          "orderDetailID": {
            "type": "integer",
            "format": "int32"
          },
          "orderID": {
            "type": "integer",
            "format": "int32"
          },
          "conDetID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "expectedQty": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "detSODID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "woNumber": {
            "type": "string",
            "nullable": true
          },
          "dirtyStockCode": {
            "type": "string",
            "nullable": true
          },
          "stockCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "workRequired": {
            "type": "string",
            "nullable": true
          },
          "colour": {
            "type": "string",
            "nullable": true
          },
          "warranty": {
            "type": "string",
            "nullable": true
          },
          "leadTime": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deliveryTerms": {
            "type": "string",
            "nullable": true
          },
          "despatchMethod": {
            "type": "string",
            "nullable": true
          },
          "deliveryDays": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "qtyReceived": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "unitPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "specialInstruction": {
            "type": "string",
            "nullable": true
          },
          "serialNumber": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ReceivedPart": {
        "type": "object",
        "properties": {
          "stockCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}